Fix use of upload images in web component + reorganize file & project folders
