+++ src/main/java/fr/lesprojetscagnottes/core/campaign/CampaignController.java
@@ -14,10 +14,10 @@ import fr.lesprojetscagnottes.core.donation.repository.DonationRepository;
 import fr.lesprojetscagnottes.core.organization.OrganizationEntity;
 import fr.lesprojetscagnottes.core.organization.OrganizationModel;
 import fr.lesprojetscagnottes.core.organization.OrganizationRepository;
-import fr.lesprojetscagnottes.core.project.ProjectEntity;
-import fr.lesprojetscagnottes.core.project.ProjectModel;
-import fr.lesprojetscagnottes.core.project.ProjectRepository;
-import fr.lesprojetscagnottes.core.project.ProjectStatus;
+import fr.lesprojetscagnottes.core.project.entity.ProjectEntity;
+import fr.lesprojetscagnottes.core.project.model.ProjectModel;
+import fr.lesprojetscagnottes.core.project.repository.ProjectRepository;
+import fr.lesprojetscagnottes.core.project.model.ProjectStatus;
 import fr.lesprojetscagnottes.core.slack.SlackClientService;
 import fr.lesprojetscagnottes.core.slack.entity.SlackTeamEntity;
 import fr.lesprojetscagnottes.core.user.UserEntity;
