+++ src/main/java/fr/lesprojetscagnottes/core/file/model/ImageUploadResponse.java
@@ -1,0 +1,13 @@
+package fr.lesprojetscagnottes.core.file.model;
+
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter(AccessLevel.PUBLIC)
+@Setter(AccessLevel.PUBLIC)
+public class ImageUploadResponse {
+
+    String path;
+
+}
