+++ src/main/java/fr/lesprojetscagnottes/core/file/repository/FileRepository.java
@@ -1,6 +1,6 @@
-package fr.lesprojetscagnottes.core.content.repository;
+package fr.lesprojetscagnottes.core.file.repository;
 
-import fr.lesprojetscagnottes.core.content.entity.FileEntity;
+import fr.lesprojetscagnottes.core.file.entity.FileEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FileRepository extends JpaRepository<FileEntity, Long> {
