+++ src/main/java/fr/lesprojetscagnottes/core/news/entity/NewsEntity.java
@@ -3,7 +3,7 @@ package fr.lesprojetscagnottes.core.news.entity;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import fr.lesprojetscagnottes.core.news.model.NewsModel;
 import fr.lesprojetscagnottes.core.organization.OrganizationEntity;
-import fr.lesprojetscagnottes.core.project.ProjectEntity;
+import fr.lesprojetscagnottes.core.project.entity.ProjectEntity;
 import fr.lesprojetscagnottes.core.user.UserEntity;
 import lombok.AccessLevel;
 import lombok.Getter;
