+++ src/main/java/fr/lesprojetscagnottes/core/organization/OrganizationController.java
@@ -31,10 +31,10 @@ import fr.lesprojetscagnottes.core.idea.IdeaRepository;
 import fr.lesprojetscagnottes.core.news.entity.NewsEntity;
 import fr.lesprojetscagnottes.core.news.model.NewsModel;
 import fr.lesprojetscagnottes.core.news.repository.NewsRepository;
-import fr.lesprojetscagnottes.core.project.ProjectEntity;
-import fr.lesprojetscagnottes.core.project.ProjectModel;
-import fr.lesprojetscagnottes.core.project.ProjectRepository;
-import fr.lesprojetscagnottes.core.project.ProjectStatus;
+import fr.lesprojetscagnottes.core.project.entity.ProjectEntity;
+import fr.lesprojetscagnottes.core.project.model.ProjectModel;
+import fr.lesprojetscagnottes.core.project.repository.ProjectRepository;
+import fr.lesprojetscagnottes.core.project.model.ProjectStatus;
 import fr.lesprojetscagnottes.core.slack.SlackClientService;
 import fr.lesprojetscagnottes.core.slack.controller.SlackController;
 import fr.lesprojetscagnottes.core.slack.entity.SlackTeamEntity;
