+++ src/main/java/fr/lesprojetscagnottes/core/project/model/ProjectStatus.java
@@ -1,4 +1,4 @@
-package fr.lesprojetscagnottes.core.project;
+package fr.lesprojetscagnottes.core.project.model;
 
 public enum ProjectStatus {
     DRAFT, IN_PROGRESS, FINISHED
